3.3 - Trabalhando com casas decimais

+++ adivinhacao.c
@@ -10,7 +10,7 @@ int main()
     int numerosecreto = 42;
     int chute;
     int tentativas = 1;
-    int pontos = 1000;
+    double pontos = 1000;
 
     while (1)
     {
@@ -44,13 +44,13 @@ int main()
 
         tentativas++;
 
-        int pontosperdidos = (chute - numerosecreto) / 2;
+        double pontosperdidos = (chute - numerosecreto) / 2.0;
         pontos = pontos - pontosperdidos;
     }
 
     printf("Fim de jogo!\n");
     printf("Você acertou em %d tentativas\n", tentativas);
-    printf("Total de pontos: %d\n", pontos);
+    printf("Total de pontos: %f\n", pontos);
 
     return 0;
 }
